refactor(examples): extract json example from example.rs
